fix pet and procedure deletions

+++ app/src/main/java/com/f4/mypet/ui/screens/procedure/show/ProcedureViewModel.kt
@@ -7,8 +7,9 @@ import com.f4.mypet.data.db.entities.Frequency
 import com.f4.mypet.data.db.entities.Procedure
 import com.f4.mypet.data.db.entities.ProcedureTitle
 import com.f4.mypet.data.db.entities.ProcedureType
+import com.f4.mypet.data.network.NetworkRepository
+import com.f4.mypet.data.network.model.NetworkResult
 import com.f4.mypet.ui.screens.procedure.FrequencyOptions
-import com.f4.mypet.util.PetDateTimeFormatter
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -19,17 +20,21 @@ import javax.inject.Inject
 
 @HiltViewModel
 class ProcedureViewModel @Inject constructor(
-    private val repository: Repository
+    private val repository: Repository,
+    private val networkRepository: NetworkRepository
 ) : ViewModel() {
     private val _procedureUiState = MutableStateFlow(
         Procedure(
             0, 0, "", 0,
-            LocalDateTime.parse("01.01.1001 00:00", PetDateTimeFormatter.dateTime),
-            "", LocalDateTime.parse("01.01.1001 00:00", PetDateTimeFormatter.dateTime),
+            LocalDateTime.now().withMinute(0),
+            "", LocalDateTime.now().withMinute(0),
             0, 0, 0
         )
     )
     val procedureUiState = _procedureUiState.asStateFlow()
+    // если null, то ошибок при удалении не было
+    private val _msg = MutableStateFlow<String?>("")
+    val msg = _msg.asStateFlow()
 
     var title = ProcedureTitle(name = "", type = -1, id = -1)
     var type = ProcedureType(name = "", id = title.type)
@@ -38,12 +43,11 @@ class ProcedureViewModel @Inject constructor(
     fun getProcedure(procedureId: Int) {
         viewModelScope.launch(Dispatchers.IO) {
             _procedureUiState.value = repository.getProcedure(procedureId)
-            title =
-                repository.getProcedureTitlesForCU().find { it.id == _procedureUiState.value.title }
-                    ?: title
-            type = repository.getProcedureTypes().find { it.id == title.type }
-                ?: type
             frequency = repository.getFrequency(_procedureUiState.value.frequencyOption)
+            title = repository.getProcedureTitlesForCU()
+                .find { it.id == _procedureUiState.value.title } ?: title
+            type = repository.getProcedureTypes()
+                .find { it.id == title.type } ?: type
 
             when (frequency.option) {
                 FrequencyOptions.Minutes.period -> frequency.frequency =
@@ -65,8 +69,13 @@ class ProcedureViewModel @Inject constructor(
 
     fun deleteProcedure(procedure: Procedure) {
         viewModelScope.launch(Dispatchers.IO) {
-            //TODO add delete of frequency
-            repository.deleteProcedure(procedure)
+             when (val response = networkRepository.removeProcedure(procedure.id)) {
+                is NetworkResult.Success -> {
+                    _msg.value = null
+                    repository.deleteProcedure(procedure)
+                }
+                is NetworkResult.Error -> { _msg.value = response.msg }
+            }
         }
     }
 }
